pyproject.toml
[project]
name = "communityeventsplatform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.5"
dependencies = [
    "asgiref==3.8.1",
    "django==5.0.7",
    "django-cors-headers==4.4.0",
    "django-filter==25.1",
    "djangorestframework==3.15.2",
    "djangorestframework-simplejwt==5.3.1",
    "drf-spectacular==0.29.0",
    "mysqlclient==2.2.4",
    "psycopg2-binary==2.9.9",
    "pyjwt==2.9.0",
    "python-dotenv==1.0.1",
    "sqlparse==0.5.1",
]
